File Upload Service/app/requirements.txt
PyYAML>=6.0
pandas>=1.3
numpy>=1.21
scikit-learn>=1.0
scipy>=1.7
Pillow>=9.0
streamlit>=1.48
